Bugfix start level race condition
As the startLevel() action sets status to PLAYING, the Game component will be rendered (see GameEngine render() method). This rendering may raise an error due to currentTrial being 'undefined'.
Solution: Make startLevel() action atomic and let it define it's a trial on the reducer, instead of setting it 'undefined', waiting for the newTrial() action to be dispatched after.